+++ lib/pvectl.rb
@@ -90,6 +90,7 @@ require_relative "pvectl/models/journal_entry"
 require_relative "pvectl/models/task_log_line"
 require_relative "pvectl/models/volume"
 require_relative "pvectl/models/dns_config"
+require_relative "pvectl/models/time_config"
 require_relative "pvectl/models/capability"
 
 # Repositories
@@ -109,6 +110,7 @@ require_relative "pvectl/repositories/disk"
 require_relative "pvectl/repositories/volume"
 require_relative "pvectl/repositories/dns"
 require_relative "pvectl/repositories/service"
+require_relative "pvectl/repositories/time_config"
 require_relative "pvectl/repositories/capabilities"
 
 # Presenters
@@ -139,6 +141,7 @@ require_relative "pvectl/presenters/disk"
 require_relative "pvectl/presenters/volume"
 require_relative "pvectl/presenters/dns_config"
 require_relative "pvectl/presenters/service"
+require_relative "pvectl/presenters/time_config"
 require_relative "pvectl/presenters/capability"
 
 # Selectors
@@ -277,6 +280,7 @@ require_relative "pvectl/commands/get/handlers/disks"
 require_relative "pvectl/commands/get/handlers/volume"
 require_relative "pvectl/commands/get/handlers/dns"
 require_relative "pvectl/commands/get/handlers/services"
+require_relative "pvectl/commands/get/handlers/time"
 require_relative "pvectl/commands/get/handlers/capabilities"
 
 # Commands - Describe